requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
pyyaml>=6.0
pyarrow>=12.0
matplotlib>=3.7

# Synthesizers
sdv>=1.0
sdmetrics>=0.14

# Diffusion (TabDDPM via Synthcity)
synthcity>=0.2.11

# Optional (for attacker / C2ST with XGBoost)
xgboost>=2.0
